+++ carp_esense_package/lib/esense.dart
@@ -6,7 +6,7 @@ import 'dart:math' as math;
 import 'package:json_annotation/json_annotation.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:flutter/services.dart';
-import 'package:esense/esense.dart';
+import 'package:esense_flutter/esense.dart';
 
 part 'esense_package.dart';
 part 'esense_domain.dart';