+++ carp_esense_package/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.1.1
+* upgrade to use the `esense_flutter` plugin instead of the deprecated `esense` plugin.
+
 ## 0.1.0
 * initial release